SYS now passes 64-bit values

+++ src/graphsdl.h
@@ -24,7 +24,7 @@
 #ifndef GRAPHSDL_INC
 #define GRAPHSDL_INC
 
-extern void get_sdl_mouse(int32 values[]);
+extern void get_sdl_mouse(int64 values[]);
 extern void warp_sdlmouse(int32 x, int32 y);
 extern void sdl_mouse_onoff(int state);
 extern void set_wintitle(char *title);
